udevng: Detect quectel device type with modem_type
modem_info.devices and modem_info.serial are the same location in memory
thanks to their "union" declaration. If we arrive at setup_quectel with
modem_info.devices set instead of .serial, modem->serial is likely a
non-null value and we execute setup_quectel_serial. setup_quectel_serial
attempts to ask udev to perform an action on the memory location where a
udev_device would be, but there's actually a GSList in .serial and
.serial.dev is just some random data that happens to be in memory. udev
understandably does not like this, fails an assertion, and ofonod
crashes.

Using modem_type seems to be the preferred option in this case.