DO NOT MERGE: Do not send_clcc if FLAG_NEED_CLIP is unset
We found a race condition on the EG25 where a CLIP would arrive and the
call would be announced, then the something else would announce the call
again.

This is unsuitable for merge upstream as it does not really solve the
underlying problem... the CLCC callback should be able to find the
announced call.